--- a/fastkmeans/__init__.py
+++ b/fastkmeans/__init__.py
@@ -1,4 +1,4 @@
 from .kmeans import FastKMeans
 
 __all__ = ["FastKMeans"]
-__version__ = "0.1.0"
+__version__ = "0.2.0"