This PR adds a Triton kernel to calculate the inner loop of the double chunked KMeans calculation.

On large enough datasets, the Triton backend is the fastest KMeans in the `speedtest.py` benchmark. It is significantly faster than FAISS GPU, 74.5% (~10 seconds) and 69.5% (~28 seconds) faster on 3M and 6.5M samples, respectively, while matching FAISS GPU in Normalized Mutual Info (NMI). It is also ~20-24% faster than Fast PyTorch KMeans with a higher NMI.

The kernel's heuristics have only been tested on 4090s and may require some tweaking for other Lovelace or Ampere GPUs.

Theoretically, it should also support AMD and Intel GPUs, although neither have been tested.

4090 Benchmarks:
<details>
  <summary>3,276,800 samples, 32,768 clusters</summary>

```
==================================================
BENCHMARK: 3276800 samples, 32768 clusters
==================================================
Generating synthetic data: 3276800 samples, 128 features, 32768 clusters...
Generated synthetic data: 3276800 samples, 128 features, 32768 clusters

=== Faiss KMeans on 3276800 samples, 32768 clusters ===
[Faiss KMeans] Done in 13.8605 seconds
[Faiss KMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9780

=== FastKMeans on 3276800 samples, 32768 clusters ===
[PyTorch FastKMeans] Done in 10.6221 seconds
[PyTorch FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9777

=== FastKMeans on 3276800 samples, 32768 clusters ===
[PyTorch FastKMeans] Done in 3.5253 seconds
[Triton FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9777

=== Fast PyTorch KMeans on 3276800 samples, 32768 clusters ===
[Fast PyTorch KMeans] Done in 4.3787 seconds
[Fast PyTorch KMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.8958
```
</details>

<details>
  <summary>6,553,600 samples, 65,536 clusters</summary>

```
==================================================
BENCHMARK: 6553600 samples, 65536 clusters
==================================================
Generating synthetic data: 6553600 samples, 128 features, 65536 clusters...
Generated synthetic data: 6553600 samples, 128 features, 65536 clusters

=== Faiss KMeans on 6553600 samples, 65536 clusters ===
[Faiss KMeans] Done in 48.7185 seconds
[Faiss KMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9792

=== FastKMeans on 6553600 samples, 65536 clusters ===
[FastKMeans] Done in 40.4036 seconds
[PyTorch FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9791

=== FastKMeans on 6553600 samples, 65536 clusters ===
[FastKMeans] Done in 12.3190 seconds
[Triton FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9792

=== Fast PyTorch KMeans on 6553600 samples, 65536 clusters ===
[Fast PyTorch KMeans] Done in 16.2451 seconds
[Fast PyTorch KMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.8209
```

</details>