@bclavie 2025-04-14
LGTM. Removed the unfair setting for Fast PyTorch Kmeans (which made it fast by vastly downsampling the data) and added an updated perf plot.